Move buildutils to scripts
