--- packages/buildutils/README.md
@@ -1,1 +1,0 @@
-# @jupyterlab/buildutils
