+++ scripts/src/build.ts
