+++ scripts/src/index.ts
@@ -1,0 +1,9 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) Jupyter Development Team.
+| Distributed under the terms of the Modified BSD License.
+|----------------------------------------------------------------------------*/
+
+export * from './build';
+export * from './ensure-package';
+export * from './get-dependency';
+export * from './utils';
